Consider amount of frames instead of buffers per Ogg page
The main reason behind this change is the situation,
when the size of the device's buffer lesser than the frame size
the library sends empty ogg pages with no data in them.

For example, with buffer size 256 which equals to about 6ms of time,
maxBuffersPerPage: 1 and frameSize: 10ms
the library sends 172 (1000 / 6) packets per second,
only 100 (1000 / 10) contain data.

Having the buffer size lesser than the frame size improve latency on streaming